refactored the new network code out of main.rs and added to new network manager module
