+++ src/network/mod.rs
@@ -3,3 +3,4 @@
 pub mod syndactyl_behaviour;
 pub mod syndactyl_p2p;
 pub mod transfer;
+pub mod manager;
